fix(setup): fix lingering temp and app files

+++ Project/MyProcControlLite/setupui.cpp
@@ -1,11 +1,15 @@
 ﻿#include "setupui.h"
 #include <w32use.hpp>
 #include <shlobj.h>
+#include <fstream>
 #include "configtool.hpp"
+#include "processhelper.h"
 #include "resource.h"
 using namespace std;
 
 extern HINSTANCE hInst;
+
+signed char __stdcall FileDeleteTreeW(std::wstring szPath, bool = false);
 
 
 int RunConfigUI(std::wstring name) {
@@ -113,10 +117,18 @@ int RunSetupUI(std::wstring name, std::wstring action, bool interactive) {
 			return 0;
 		}
 		if (action == L"uninstall") {
+			BOOL keep = false;
 			if (interactive) {
-				TaskDialog(NULL, hInst, L"Uninstall MyProcControl (Lite)",
-					L"Are you sure you want to uninstall this product?", name.c_str(),
-					TDCBF_YES_BUTTON | TDCBF_CANCEL_BUTTON, MAKEINTRESOURCEW(IDI_ICON_APP), &btn);
+				TASKDIALOGCONFIG c{};
+				c.cbSize = sizeof(c);
+				c.dwCommonButtons = TDCBF_YES_BUTTON | TDCBF_CANCEL_BUTTON;
+				c.hMainIcon = LoadIconW(hInst, MAKEINTRESOURCEW(IDI_ICON_APP));
+				c.pszVerificationText = L"&Keep the application executable file?";
+				c.pszWindowTitle = L"Uninstall MyProcControl (Lite)";
+				c.pszMainInstruction = L"Are you sure you want to uninstall this product?";
+				c.pszContent = name.c_str();
+				c.dwFlags = TDF_USE_HICON_MAIN;
+				TaskDialogIndirect(&c, &btn, NULL, &keep);
 				if (btn != IDYES) return ERROR_CANCELLED;
 			}
 			Service myService = scm.get(name);
@@ -128,13 +140,55 @@ int RunSetupUI(std::wstring name, std::wstring action, bool interactive) {
 			if (myService.status() == SERVICE_PAUSED) {
 				if (!myService.stop()) throw w32oop::exceptions::system_exception("Failed to stop service.");
 			}
-			RegistryKey uninstall(HKEY_LOCAL_MACHINE, L"SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Uninstall");
-			uninstall.delete_key(L"Service_" + name);
-				
+			try {
+				RegistryKey uninstall(HKEY_LOCAL_MACHINE, L"SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Uninstall");
+				if (uninstall.exists(L"Service_" + name)) try {
+					auto p = uninstall.open(L"Service_" + name);
+					auto tmp = p.get_value<wstring>(L"MyData:TempDir");
+					if (!tmp.empty()) FileDeleteTreeW(tmp + L"\\");
+				} catch (...) {
+					MessageBoxW(0, ErrorChecker().message().c_str(), 0, MB_ICONHAND);
+				}
+				if (uninstall.exists(L"Service_" + name)) uninstall.delete_key(L"Service_" + name);
+			}
+			catch (...) {}
+
 			if (interactive) TaskDialog(NULL, NULL, L"MyProcControl (Lite) Setup",
 				L"The product has been successfully removed from your computer.",
 				L"Click OK to dismiss.",
 				TDCBF_OK_BUTTON | TDCBF_CANCEL_BUTTON, TD_INFORMATION_ICON, NULL);
+
+			if (!keep) {
+				WCHAR buf[260]{};
+				GetTempPathW(260, buf);
+				wstring f = buf;
+				f += L"\\";
+				f += GenerateUUIDW();
+				f += L".sct";
+				auto appPath = make_shared<WCHAR[]>(32768);
+				GetModuleFileNameW(NULL, appPath.get(), 32768);
+				ofstream fp(f, ios::out | ios::binary);
+				wstring sApp = appPath.get();
+				w32oop::util::str::operations::replace(sApp, L"\\", L"\\\\");
+				wstring cont = L"<?XML version=\"1.0\"?><scriptlet><registration classid=\"{" + GenerateUUIDW() +
+					L"}\"><script language=\"JScript\"><![CDATA[\r\nvar i=0,f=\"" + sApp + L"\",a=new ActiveXObject(\""
+					"Scripting.FileSystemObject\");for(;i<9999999;++i){try{if(!a.FileExists(f))break;a."
+					"DeleteFile(f,true);break}catch(e){}}\r\n]]>\r\n</script></registration></scriptlet>";
+				fp.write("\xff\xfe", 2);
+				fp.write((PCSTR)cont.data(), cont.size() * sizeof(decltype(cont)::value_type));
+				fp.close();
+				STARTUPINFOW si{ sizeof(si) }; PROCESS_INFORMATION pi{};
+				WCHAR s32[260]{}; GetSystemDirectoryW(s32, 260);
+				wstring cmd = format(L"regsvr32 /u /n /s /i:\"{}\" \"{}\\scrobj.dll\"", f, s32);
+				if (!CreateProcessW((s32 + L"/regsvr32.exe"s).c_str(), cmd.data(), 0, 0, 0, 0, 0, 0, &si, &pi)) {
+					if (interactive) MessageBoxW(NULL, (L"Cannot Delete Remaining File! Please manually delete "s +
+						appPath.get()).c_str(), NULL, MB_ICONERROR);
+				}
+				else {
+					CloseHandle(pi.hProcess);
+					CloseHandle(pi.hThread);
+				}
+			}
 			return 0;
 		}
 		if (interactive) MessageBoxW(NULL, L"Unknown action type", L"Error", MB_ICONERROR);
@@ -149,4 +203,43 @@ int RunSetupUI(std::wstring name, std::wstring action, bool interactive) {
 		return int((!!err) ? err : 1);
 	}
 }
+
+
+signed char __stdcall FileDeleteTreeW(std::wstring szPath, bool __internal_is_first_call__) {
+	signed char status = 0;
+	if (__internal_is_first_call__) {
+		if (szPath.find(L"/") != wstring::npos) w32oop::util::str::operations::replace(szPath, L"/", L"\\");
+		if (!szPath.ends_with(L"\\")) szPath += L"\\";
+	}
+	WIN32_FIND_DATAW findd{};
+	HANDLE hFind = FindFirstFileW((szPath + L"*").c_str(), &findd);
+	if (!hFind || hFind == INVALID_HANDLE_VALUE) {
+		status = -1;
+		return status;
+	}
+	do {
+		if (wcscmp(findd.cFileName, L".") == 0 ||
+			wcscmp(findd.cFileName, L"..") == 0) continue;
+		wstring wstrFileName;
+		wstrFileName.assign(szPath);
+		//wstrFileName.append(L"\\");
+		wstrFileName.append(findd.cFileName);
+		if (findd.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) {
+			wstrFileName.append(L"\\");
+			auto result0 = FileDeleteTreeW(wstrFileName, false);
+			if (result0 != 0) status = (status == -1) ? -1 : 1;
+		}
+		else {
+			if (findd.dwFileAttributes & FILE_ATTRIBUTE_READONLY)
+				SetFileAttributesW(wstrFileName.c_str(), FILE_ATTRIBUTE_NORMAL);
+			if (!DeleteFileW(wstrFileName.c_str()))
+				status = (status == -1) ? -1 : 1;
+		}
+	} while (FindNextFileW(hFind, &findd));
+	FindClose(hFind);
+	if (!RemoveDirectoryW(szPath.c_str()))
+		status = (status == -1) ? -1 : 1;
+	return status;
+}
+
